Moved trait tests to new file
