+++ tee/src/tests/mod.rs
@@ -13,6 +13,7 @@
 
 // You should have received a copy of the GNU General Public License
 // along with Ternoa.  If not, see <http://www.gnu.org/licenses/>.
-
 mod extrinsics;
+mod traits;
+
 pub mod mock;
